feat: add animation sidebar
