+++ client/src/app/messages/_components/sidebar/option/savedMessage.tsx
@@ -4,6 +4,8 @@ import React from 'react';
 interface MyProfileProps {
   showSavedMessages: boolean;
   setShowSavedMessages: React.Dispatch<React.SetStateAction<boolean>>;
+  isShowOptionProfile: boolean;
+  setIShowOptionProfile: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
 const savedMessages = [
@@ -19,15 +21,18 @@ const savedMessages = [
   },
 ];
 
-const SavedMessages: React.FC<MyProfileProps> = ({ showSavedMessages, setShowSavedMessages }) => {
+const SavedMessages: React.FC<MyProfileProps> = ({ showSavedMessages, setShowSavedMessages,setIShowOptionProfile }) => {
   if (!showSavedMessages) return null;
 
   return (
-    <aside className="min-w-1/4 w-full md:w-1/4 h-screen bg-white shadow-md overflow-y-auto fixed sidebar z-50">
+    <aside className="bg-white shadow-md overflow-y-auto fixed sidebar z-20 w-full h-full">
       <div className="flex justify-between px-6 py-4 items-center border-b">
         <h2 className="text-lg font-semibold text-gray-700">Saved Messages</h2>
         <ArrowLeft
-          onClick={() => setShowSavedMessages(false)}
+          onClick={() => {
+            setShowSavedMessages(false)
+            setIShowOptionProfile(false)
+          }}
           className="cursor-pointer w-6 h-6 text-gray-600 hover:text-gray-800 transition-all duration-300"
         />
       </div>
